Adapting maintenance class, adding to demo, updating UML

+++ Facility/src/maintenance/Issue.java
@@ -4,10 +4,10 @@ public class Issue {
 
     //default access is package private so that subclasses may use them directly
     String details;
-    double time; // Time in hours, can be decimal
-    double cost;
+    long time; // Time in hours, can be decimal
+    long cost;
 
-    public Issue(String details, Double time, Double cost) {
+    public Issue(String details, long time, long cost) {
         this.details  = details;
         this.time     = time;
         this.cost     = cost;
@@ -23,11 +23,11 @@ public class Issue {
 
     public void setDetails(String details) { this.details = details; }
 
-    public double getTime() { return time; }
+    public long getTime() { return time; }
 
-    public void setTime(double time) { this.time = time; }
+    public void setTime(long time) { this.time = time; }
 
-    public double getCost() { return cost; }
+    public long getCost() { return cost; }
 
-    public void setCost(double cost) { this.cost = cost; }
+    public void setCost(long cost) { this.cost = cost; }
 }
